Move isVideoStationPage.ts to lib
